==> Models/02_Naive_Bayes.ipynb <==
import pandas as pd

df = pd.read_csv('../Data/reviews_training_26000.csv')
val = pd.read_csv('../Data/reviews_test_4000.csv')

df = df[['review','sentiment']]
val = val[['review','sentiment']]

df['sentiment'] = df['sentiment'].map({'positive':1,'negative':0})

### EDA

df['sentiment'].value_counts() #More Negative than positive

val['sentiment'].value_counts() #Inverse in Test

### Model Building

train = df[:int(0.8*len(df))]
test = df[int(0.8*len(df)):]

from sklearn.feature_extraction.text import CountVectorizer
vectorizer = CountVectorizer()
train_bow = vectorizer.fit_transform(train['review'])
test_bow = vectorizer.transform(test['review'])

from sklearn.naive_bayes import MultinomialNB
model = MultinomialNB()
model.fit(train_bow, train['sentiment'])

predictions = model.predict(test_bow)

from sklearn.metrics import accuracy_score
accuracy = accuracy_score(test['sentiment'], predictions)
print("Accuracy:", accuracy)

val_predictions = model.predict(vectorizer.transform(val['review']))

val_predictions

val['sentiment_vector'] = val['sentiment'].map({'positive':1,'negative':0})

accuracy_score(val['sentiment_vector'],val_predictions)

from joblib import dump

dump(vectorizer,'Artifacts/count_vectorizer.joblib')
dump(model,'Artifacts/model.joblib')
