==> Models/02_Naive_Bayes_Testing.ipynb <==
from joblib import load 

cv = load('Artifacts/count_vectorizer.joblib')
model = load('Artifacts/model.joblib')

test_data = ["terrible movie, don't watch","dramatic movie please watch"]

test_data = 'hello'

transform_data = cv.transform(test_data)

predictions = model.predict(transform_data)

predictions
